V7: list all security groups (`security-groups`)
We pulled out a helper method to build a list of spaces associated with
the security group.

We also created a global test which unbinds **all** global security groups in
order to test the empty case. This may be overkill. Watch out: to avoid
race conditions, make sure in the tests' setup that the `helper.CF()`
commands have finished (exited).

In an unrelated change, we clarified the message for resources that have
not been configured in the API router. The error now reads:

```
no resource exists with the name xxxx, did you add it to 'api/cloudcontroller/ccv3/internal/api_routes.go'?
```

Reid says: This error message should never be seen by a CLI user, so
it's okay to leave breadcrumbs for the developer to fix it.

[finishes #168655219](https://www.pivotaltracker.com/story/show/168655219)

Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>