Refactor TargetCF method to include root info
The `TargetCF` method will query the API root while it is targeting CF,
but previously did not return this information to the caller. We would
then need to issue a second query to the API root. This commit refactors
the method to return the information from the API root, eliminating the
need for a duplicate call.

Authored-by: Belinda Liu <bliu@pivotal.io>