🐞 security-group cleanup works for V7
Previously `bin/cleanup-integration` was not able to delete the leftover
security-groups because the formatting of the `cf security-groups`
command changed in V7. Specifically, the first column is now the name
of the security group instead of the second column, which was the case
for V6 (the first column was the index).

fixes:
```
Deleting security group <all> as admin
Security group <all> does not exist
OK
```

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>