v7: cf oauth-token command does not use v2 endpoints
We chose to make use of current v7 access token behavior, which will
only request a new token if the current access token is invalid or about
to expire. This is a change from the v6 behavior, which would request a
new token every single time this command was run.

[#171284099]

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>