Force UAA call in oauth client test
A test in the oauth client suite was relying on the v6 behavior of `cf
oauth-token`, which would always request a new token from UAA. We have
changed this behavior in v7 to only request a new token when needed. To
ensure this test still works, this commit adds an invalid token to force
a UAA call to be made.

[#171284099]

Authored-by: Belinda Liu <bliu@pivotal.io>