🐞 `StagePackage()` rigorously checks app↔︎package
StagePackage() ensures that a given package belongs to the given
application; however, the code had a subtle bug where it did not check
properly; it merely checked that the application had packages (but not
if they were the correct packages).

We now check to make sure the package is the correct one.

[#166524701]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>