v7: Serialize non-string env var groups as strings
When an env var has a value that is not a string (e.g. a number or a
JSON object), we now convert it to a string before sending it to the
API since the V3 API does not accept non-string values).

In doing this, we fixed bugs where the CLI would return an error if
trying to set or get env var groups that contained JSON object values.

[Finishes #172094353]
[Finishes #172094093]
[Finishes #172157096]

Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>