V7: Add --space and --organization flags to the copy-source command
As part of this commit, we changed the behavior of the
AppStager.StageAndStart function to take in a space, rather than
assuming that the app to be staged and started was in the currently
targeted space. This was necessary as the targeted app may be in a
different space.

We also removed the Config as part of the AppStager interface as it was
only being used to determine the space of the app to be staged and
started, which is no longer necessary

These changes necessitated adjustments in restage_command.go, which
appears to be the only other place using the AppStager.

[finishes #172070526]

Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>
Co-authored-by: Teal Stannard <tstannard@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>