Backfill integration test for `passwd`
- The command remains a V6 command, not yet V7
- We tested the following:
  - successful change
  - new & verification passwords don't match
  - wrong current password
  - current & new passwords are the same
  - zero-length new & verification password
  - not logged in

We made a change to `readPassword()` to use the `ui.scanner`. This
fixed a fairly odd error that only occurred when more than one password
was read from stdin, which is the case for the `passwd` command.
Specifically, the original version of `readPassword()` instantiated a
standard-sized 4k buffer to read the password, which had the side effect
of reading _all_ the passwords (current, new, and verification) in one
swoop, discarding the new & verification passwords, causing all tests to
fail with `Password must be at least 1 characters in length`.

The UI struct was previously enhanced to include a scanner, but wasn't
used by `AskForPassword()`, only by `Ask()`.

[#172137337]

Co-authored-by: Weyman Fung <wfung@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>