V6: `passwd` displays the logged-in user
We display the username whose password is being changed (always the
logged-in user). This is to conform with the CLI Styleguide.

Always display the username even when the new & verification passwords
don't match. Prior that was the only case where it didn't display its
intention, and that was because that particular check was performed on
the client side. Now we always display the intention unless the user is
not logged in.

```diff
- Changing password...
+ Changing password for user <logged-in-user>...
```

[#172137337]

Co-authored-by: Weyman Fung <wfung@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>