Make shared logs integration test support v6 and v7
These tests (integration/shared/isolated/logs_command_test) mock the
HTTP requests they make, including those to CAPI. Some of these requests
are made while doing 'cf login' in the test setup.

We recently ported the login command to v7, and now there is a small
difference in the HTTP requests that the command makes between v6 and
v7: in v7, when fetching spaces for an org, we include the `order_by`
query parameter. In v6, we don't.

Since these tests are shared, they began failing when run with v7, since
the call with the query parameter was not mocked out. To allow these
tests to work with both v6 and v7, we add a mock handler for the
/v3/spaces endpoint both with and without the `order_by`.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>