Split shared oauth_client_test into v6/v7 tests
We recently changed the behavior of the v7 login command to error out on
the deprecated pattern of manually specifying oauth client credentials
in the config file. These tests, which were in shared, were testing this
pattern.

Splitting them into v6/v7-specific versions so that in the v7 tests, we
can expect to see the new error message in this case.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>