V7: `auth` disallow manually-added client creds in config file
We don't like client creds in the config file; they're a security risk.
We allow them to be passed in from the command line when needed.

In V6, we gave a deprecation warning when client credentials were in the
config file, but we would allow it (the command would succeed).

In V7, we take a harder line: they're no longer allowed. This change has
already been implemented for `cf login`.

[#172316246]

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>