v7: Try to speed up domains integration tests
The `domains` command integration tests are among our slowest in the
pipeline. This is an attempt to speed them up.

These tests were creating some test orgs in BeforeEaches, and in those
orgs were creating some spaces and private domains. We don't need to
delete the spaces and private domains individually before deleting the
org, since the API will do that for us.

Also, use `helpers.QuickDeleteOrg` instead of waiting for `cf
delete-org` to complete. This just sends an async delete request without
polling until it's done (which is okay for our tests).

Authored-by: Reid Mitchell <rmitchell@pivotal.io>