V7: manifests: allow both `disk_quota` and `disk-quota`
- `disk_quota` (underscore) is the old way; we maintain
  backwards-compatibility
- `disk-quota` (hyphen) is our new convention

The reason for this inconsistency is historical: back before server-side
manifests were supported, when manifest parsing was done fully on the
CLI side, the CLI expected an underscore for `disk_quota` but a hyphen
for all other multi-word keys (not clear why). When server-side
manifests were introduced, the CAPI endpoints actually support both
hyphens and underscores interchangeably, to provide some flexibility
here, but the manifest spec still calls for an underscore in
`disk_quota` and hyphens in other keys. We are not changing the spec,
but simply allowing the CLI to parse manifests that use hyphens
everywhere for consistency. The API side should not need to change.

[#167953802]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>