WIP: refactored curl-command has basic functionality
* created a new requester function: takes in a URL, raw bytes from the
request body json, and returns raw bytes which can be directly returned
to the curl user
* Wrote a command layer that covers the `-X`, `-d`, and `-H` flags.
* Unit testing the current implemetnation of the command layer is
blocked on the `cloudControllerClient` being a private field on the base
command so it can't be mocked out. This might mean we should follow
the patterns and add an actor and client layer
* Other flags have not started (but see comment in requester for
implementation ideas). Additionally, the `-v` flag prints output twice
(once from `-v` responses and once from `cmd.UI.DisplayText`)

[#172322551]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>