v7: Improve performance of `cf apps`
The `cf apps` command was noticeably slower in v7 than in v6. For a
space with about 20 apps, it took 1-2s in v6 and 8-10s in v7. This
change makes some simple improvements to bring it down to about 4-5s.

Performance improvements:

- Previously, for each process in each app in the targeted space, we
were making a request for the process' sidecars. We weren't displaying
this information in this command, so we refactored things to not make
this request. We need sidecar data for the `cf app` command, which
prints a more detailed summary of a single app.

- Previously, we were going through the apps in the space and, for each
one, making a request like `GET /v3/apps/:guid/processes`. Now, we
collect all app guids that we care about and make a request outside of a
loop like `GET /v3/processes?app_guids=guid1,guid2,guid3,...`. This
added a little more logic to map processes back to their parent apps,
but it was worth it.

- Previously, when fetching mapped routes, we were making a request for
each route's space so we could retrieve the space name. We aren't
printing the space name for this command (that's used in `cf routes`),
so we refactored things to not share this logic between commands (and
thus not make these unnecessary requests for each route in `cf apps`).

Other changes & refactors:

- Moved a few struct definitions (`Route`, `Metadata`, etc.) out of the
ccv3/v7action packages and into the new shared resources package.
Deleted the no-longer-needed `ccv3.Route`/`v7action.Route`/etc.

[#172461503]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>