V7: Don't cache API Version in the config file
The `api` and `login` commands no longer respect the APIVersion in the
config file; instead, they write the APIVersion anew.

This sets the stage for setting a minimum API level for the CLI, for
when the APIVersion is cached, and V6 and V7 CLIs are used
interchangeably, the v2 API version can be mistakenly persisted instead
of the v3 API version.

The actor method no longer `short-circuits` when detecting a matching
API endpoint and skipSSLValidation directive in the configuration file.

[#172208936]

Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>