V7: **login/auth** displays minimum support CC API warning
The warning only appears when the CC API version does not meet the
minimum (3.84.0). If the minimum is met, no warning appears.

This is a warning, not an error. Users may continue using the V7 CLI
when the minimum API version is not met.

The warning **only** appears on `login` & `auth`. We don't want to
overwhelm the user with this warning on every command.

Developer Notes:

- we skipped integration tests, for there was no easy way to set the
API version of the cloud controller
- shared code is in `command/v7/shared/version_checker*.go`

[finishes #172208936]

Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>