V7: **login/auth** does not error if indeterminate CC API version
If **login/auth** cannot determine the CC API version, or cannot compare
it to the minimum supported version, it prints a warning to that effect
and continues operation.

Previously it would error, and prevent the user from authenticating or
performing any operations requiring authentication. This restriction was
too draconian for the story, whose intent was to merely warn the user if
there was something they should be worried about (a too-old CC API); its
intent was never to prevent the user from getting their work done.

[finishes #172208936]

Authored-by: Brian Cunnie <bcunnie@pivotal.io>