V7: `add-network-policy`'s dest is arg, not flag
This also applies to `remove-network-policy`.

The earlier incarnation of `{add,remove}-network-policy` had a required
flag, `--destination-app`. But our style guide suggests that required
args not have flags, so it's no longer a flag, it's the second argument.

This commit brings the two commands into conformance with our style
guide.

[#172617004]