🐞 V7: `reset-org-default-isolation-segment` TIP
`reset-org-default-isolation-segment` now gives a "tip" to restart
applications for the change to take effect. This is similar to other
commands.

We took the liberty of tweaking the tip; specifically, we replace the
words "platform's default isolation segment" with "platform's default".
The platform does not have a default isolation segment; it has the
shared segment, which is not isolated.

[finishes #172665289]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>