WIP: implement v8 version of api command
- 1st attempt: Non-shared code approach of just creating a v8 command list, base command, actor
that are clones of the v7 version. I believe this is what v8 could look
like if these interfaces diverge in the future and stop sharing code.
This is probably not what we would want going forward as changes to
cloned classes would need to be manually applied to future versions

[#172439514](https://www.pivotaltracker.com/story/show/172439514)