v7: Improve performance of `apps` command
We recently added an `app_guids` filter to the `GET /v3/routes` endpoint
on the API, which will return routes mapped to the specified apps. This
commit updates the `apps` command to make use of that new filter.

Previously, we would go through all the apps in the table and fetch the
mapped routes for each one (with `GET /v3/apps/:guid/routes`).

Now, we make a single* request for all the routes that are mapped to the
set of apps we're displaying. (*may not actually be a single request, if
there are multiple pages of routes returned, but at least it doesn't
scale with the number of apps).

[Finishes #172625875]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Teal Stannard <tstannard@pivotal.io>