v7: filter orgs in login command if -s flag is provided without -o flag
Previously, if a user provided a space but had access to multiple orgs,
the prompt would show all orgs and then error if the user selected an
org that did not contain the requested space.

Now the list of orgs will be filtered to those containing the space, and
the user gains the benefit of auto targetting if only one org contains
the space.

This does add API calls to check for the space's presence in orgs the
user has access to, but it will only run through that filtering if the
space flag is set without an org flag.

[finishes #172414819]

Authored-by: Sarah Weinstein <sweinstein@pivotal.io>