v7: Add `--router-group` flag to `create-shared-domain`
This allows the creation of TCP domains.

As part of this change, we refactored the way we are making requests to
the routing API in v7. Previously, we were using a method on the
`api/cloudcontroller/ccv3` client, like our normal CAPI requests.

Now, we are using a dedicated routing API client, located at
`api/router`, to make these requests. This client is able to use the
same connection-wrapping pattern we have for the CAPI and UAA clients,
which allows automatic token refreshing (among other things).

[Finishes #172510293]
[Finishes #172686580]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Belinda Liu <bliu@pivotal.io>