v7: cf label commands can accept a route with a port
This commit updates the route parsing logic to handle routes with ports.
The `cf set-label` and `cf unset-label` commands should now be able to
retrieve and update routes that look like `domain:port`.

We included an integration test for this new behavior, but left it
commented out because it relies on creating a route with a port through
the CLI which is not yet implemented. Once the `--port` flag is added to
the `create-route` command in a later story, this test should be
uncommented.

[#172665133]

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>