v7: Remove port from domain name in `routes`
This fixes a bug in the output of the `routes` command:

When a route has a port, we were showing its domain in the table like:

```
domain.com:1234
```

This is because we are extracting the domain name from the route's URL
field. We were stripping off the host and path, but not the port. Now,
we strip off the port as well.

[Finishes #172922420]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>