test: fake hostnames should be generated in lower case
Since
https://github.com/cloudfoundry/cli/commit/c83198772aaff96fa3d18df495af333b5b2af447
was merged there have been CI failures. That change (a merge of
https://github.com/cloudfoundry/cli/pull/1853) means that routes
would be converted to lower case. A number of routes and labels tests
started to fail after this because the output from commands no longer
matched the upper case inputs. It seems pragmatic that fake hostnames
should in general be in lower case so that tests are simpler.