v7: Do not lowercase hostnames in `map-route`
This undoes the changes introduced in
c83198772aaff96fa3d18df495af333b5b2af447.

Those changes caused a CATs failure. See the linked story below for more
investigation notes, but the gist is:

- In v6, we made the change to lowercase hostnames in `map-route`
because we were already doing it for `push`, and we felt we should be
consistent
- In v7, `push` does not lowercase hostnames, so it's strange to do it
only for `map-route` like this
- There are other ways to specify hostnames (e.g. in manifest files),
that this logic did not account for, so I think removing it entirely and
passing user input directly to the API in all cases is better.
- There is a C2C networking bug that means hostnames on internal routes
are case sensitive. That's what was actually causing the failure. More
on that here: https://github.com/cloudfoundry/cf-networking-release/issues/78

[#173019147](https://www.pivotaltracker.com/story/show/173019147)
[#169941315](https://www.pivotaltracker.com/story/show/169941315)

Authored-by: Reid Mitchell <rmitchell@pivotal.io>