Create new router group in create-route integration tests
Previously, we were using `FindOrCreateTCPRouterGroup`, which may end up
finding an existing router group. When we then create a route using that
router group on port 1024, the port may already be in use when the test
runs again. This was causing a failure like:

```
[15] CMD> cf create-route integration-TCP-DOMAIN-0f24a3d6-2876-40ee-6a83-9caa32a7d885.com --port 1024
[15] OUT: Creating route integration-TCP-DOMAIN-0f24a3d6-2876-40ee-6a83-9caa32a7d885.com:1024 for org INTEGRATION-ORG-3fc34b70-10b7-44e2-6305-43e530f9f5bb / space INTEGRATION-SPACE-72b9cab6-c781-49e8-68bc-5a0de5cabd01 as admin...
[15] ERR: Port '1024' is not available. Try a different port or use a different domain.
[15] OUT: FAILED
```

Creating/deleting a router group every time should let us avoid this
port-unavailable problem.

[finishes #172951150]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>