Do not send `ports` filter to API when not necessary
In our recent work to support TCP routes in v7/v8, we started sending
the `ports` filter to the API alongside `hosts`/`paths`/etc to uniquely
identify a route. This was harmless, since passing `ports=0` when the
`--port` flag was unspecified just excluded routes with ports that we
didn't mean to include anyway.

However, doing this mean that all commands that fetched a route by
attributes (e.g.  `map-route`, `unmap-route`, `delete-route`,
`set-label`, `unset-label`) now had a strict dependency on the edge
version of CAPI, which supports the `ports` filter on the `/v3/routes`
endpoint. We did not intend for these commands to only work against the
very latest (unreleased) CAPI version.

With this change, we only send the `ports` filter when the `--port` flag
is explicitly passed. This means that in most cases, we won't require
the CAPI to be super new (and support the `ports` filter).

[finishes #173165479]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>