Only send `ports` filter for TCP domains
A follow-up to 86e304fd05a23a28bd55c9cd5773cef630eefbfa.

We want to only send this filter when we know the API supports it. This
is true for TCP domains. Previously, we were excluding the ports filter
when the value was 0 or "" (empty string), but this did not work for the
`map-route` case where we want to ensure a new route is created with a
random port when the `--port` flag is not given.

[finishes #173165479]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>