Fix flake in download-droplet unit test
We observed that, with a certain random test ordering, this test would
fail because the droplet was not written to the expected file. It seems
like the CWD was not getting set consistently?

Could reproduce this failure consistently by running the following (note
the seed):

```
ginkgo -r -randomizeAllSpecs -requireSuite -randomizeSuites -p -skipPackage integration,cf/ssh,plugin,cf/actors/plugin,cf/commands/plugin,cf/actors/plugin --seed 1592263733
```

And making this change gets the test to pass with that seed.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>