Skip TCP domains when finding default domain in integration tests
We recently began seeing failures when a TCP route (a route using a TCP
domain) was being used unexpectedly. I believe it is because this
DefaultSharedDomain helper was not excluding domains with protocol
"tcp". Now, we exclude them, just like we do for internal domains, so
that the "default domain" it finds will never be a TCP domain.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>