Double encode name parameters in requests to CAPI
The v3 API requires that the comma character is double encoded when used
in a single query parameter. Previously, the CLI would only encode the
query request once it was fully formed. This was sufficient for v2, but
with the v3 API this would cause issues when using a resource that had
commas in its name.

This commit now encodes query parameters with the
`names` filter before encoding the entire query. This ensures that
resource names that may contain commas are properly double encoded.

[Github issue #1938](https://github.com/cloudfoundry/cli/issues/1938)
[#172891571](https://www.pivotaltracker.com/story/show/172891571)

Authored-by: Belinda Liu <bliu@pivotal.io>