Use correct authorization endpoint when logging in
Our newly-refactored `login` command works by hitting CC's `/` endpoint
and taking the authorization endpoint from the response (and then using
it to do authentication). There are two similar endpoints returned, one
under `uaa` and one under `login`. These endpoints are always the same
in standard cf-deployment installations, but they are not necessarily
the same. In BlueMix, for example, these endpoints are different.

We were incorrectly reading the `uaa` endpoint instead of the `login`
one. This fixes that.

Note that this is very similar to another regression that we fixed in v6
a few months ago. Somehow in the process of passing the code and work
between v6 and v7, we ended up making the same mistake again. Links
related to that regression are below.

In the process of fixing this, I found that we are writing an
authorization endpoint to the config file, but are apparently never
reading the value we write there. For consistency, I updated that code
to write the correct endpoint to the file (`login`, not `uaa`), but we
might consider removing that entirely someday.

For this bug:

GitHub issue: https://github.com/cloudfoundry/cli/issues/1964
Tracker story: [#173515076](https://www.pivotaltracker.com/story/show/173515076)

For the previous, similar regression:

GitHub issue: https://github.com/cloudfoundry/cli/issues/1820
Tracker story: [#169207141](https://www.pivotaltracker.com/n/projects/892938/stories/169207141)

Authored-by: Reid Mitchell <rmitchell@pivotal.io>