Allow commands to specify usage and examples with methods
Our existing pattern restricted us to writing usage and example text in
a single struct tag, which got very hard to maintain for multiple lines
of text (We would have to write newlines and indentation manually, and
it was hard to read in the end).

With this change, commands can opt to specify their usage and/or example
text by implementing `Usage()`/`Examples()` methods. The existing
pattern still works, but this lets us gradually move commands to this
easier way of doing it.

Also convert `map-route` to do it in this new way, as an example.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>