Simplify CCV3 internal router
- Now, we don't need to hit `/v3` before every command to populate the
router. Instead, we have just written the resource names directly into
the routes file.
- Refactor the routes file to be a map keyed by route name, instead of
an array of routes that gets later converted into a map on the fly.
- The `GetInfo()` method now does not return `ccv3.ResourceLinks`,
because we don't hit `/v3`.

Slack discussion where we asked if anyone knew why it was so complicated
(and nobody really knew):
https://cloudfoundry.slack.com/archives/C032824SM/p1594316115396400

Co-authored-by: Merric De Launey <mdelauney@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>