Avoid hitting `/login` endpoint in every command
Before this change, every command would reach out to the authorization
server's `/login` endpoint when initializing the UAA client. It did this
in order to figure out the UAA URL, which is needed to make UAA
requests.

To prevent this request from being made every time, and to simplify the
code, we now store the UAA endpoint in the config file (along with the
authorization endpoint) upon successful login/target of a new CF. Then
we just read the URL from there to initialize the UAA client in
subsequent commands.

This simplicity comes at the expense of some flexibility: for example,
if the UAA endpoint were to change since the user last logged in (e.g.
if the platform was re-deployed with different configuration), the user
would need to log out and log back in to retrieve the new UAA URL.
However, we are assuming that this is not a common occurrence, since
many clients/scripts rely on knowing the UAA URL, so we imagine that
this sort of configuration change would not be done lightly.

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Merric De Launey <mdelauney@pivotal.io>