Split auth integration tests into v6/v7 folders
These tests lived under `shared`, but they would fail when run with the
custom client credentials env vars set. This is because in v7, we made
it an error to write custom client creds to your config file, whereas it
was just a warning in v6.

We still had shared tests expecting it to only be a warning, which was
an oversight that we didn't notice because the custom client creds env
vars were not being set in CI.

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Merric De Launey <mdelauney@pivotal.io>