Print empty arrays correctly in verbose output
Prior to this change, empty JSON arrays were being printed as `null`.
This is due to the way our JSON-sanitization logic was allocating data
for lists. We had `var list []interface{}`, which will not allocate
anything until elements get appended to it (and in the empty case, none
would be appended). So this ended up serialized as `null`.

Changing the line to `list := make([]interface{}, len(v))` seems to
allocate memory differently, even in the empty case, such that it gets
serialized as `[]`.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>