Refactor `cf curl` to match v7 architecture
All flags are consistent with v6 implementation.

Unlike in v6, if -i and --output flags are both provided, the returned
headers will be output to the specified file, rather than to stdout. We
believe this is more in line with expected user intent.

`cf curl` will no longer attempt to refresh a users token if the
`Authorization` header is manually provided. We made this decision based
on the edge case that separate users could be logged in and using curl.
Prior art from [this
story](https://www.pivotaltracker.com/story/show/162431844) seems to
indicate that this was an unintentional feature in v6.

We removed the integration tests looking for a `Connection` header to
have been set, as the default in HTTP requests is to have `Connection`
set to `close`, which was being hard coded in v6.

[finishes #172322551]

Co-authored-by: Teal Stannard <tstannard@pivotal.io>
Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>