Read log cache endpoint from config, not client, in commands
After a recent refactor, we are now supposed to read all component URLs
from the config rather than from the CC client struct (which is no
longer dynamically populated at the start of each command). We caught
most of these cases, but a handful of commands that needed the log cache
endpoint were still reading it from the client. This commit switches
them to use the config.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>