Update proxy-connection integration test
This test aimed to run some command with the `https_proxy` env var set
to an invalid value, to exercise the case where the proxy is not set up
correctly and you get an error.

We were using `cf orgs` in this test before, which was fine because the
command would always make some initial "bootstrapping" requests to `GET
/`, etc. These bootstrapping requests would fail with the right error.

In a recent refactor, we got rid of the bootstrapping requests, so the
`cf orgs` command was not triggering the right error. Instead, we were
getting to the point of checking if the user was logged in, and
returning a "not logged in" error instead.

Fixing this by switching the command in use from `orgs` to `login`,
since the `login` command doesn't try to check if the user is logged in
before hitting the API. Thus we error out in the right way (the API
request fails due to the proxy).

Authored-by: Reid Mitchell <rmitchell@pivotal.io>