Fix verbose-flag integration test
This test was doing a `cf target -o <org> -v` and then looking for a
`Set-Cookie:` header in the output (to test the redaction of the cookie
value). However, this header only appears when we hit the UAA API, and
after a recent refactor, the `target` command does not hit UAA anymore.

Switched this test to use `auth`, like the test above it in this file,
so that we can hit UAA and properly test the appearance (and redaction)
of the `Set-Cookie` header.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>