Fix display issue regarding multi-buildpack
In the case of multi-buildpack, it seems only one buildpack populates
the `detect_output` field. However, after talking with the original
authors of the PR, it seems that `buildpack_name` in the default
provided buildpacks in this case. This gives us a better separation
between the `buildpack names:` and `buildpacks` fields.

[#173716617](https://www.pivotaltracker.com/story/show/173716617)

Co-authored-by: Sebastian Vidrio <svidrio@pivotal.io>