`cf apply-manifest` displays diff
This uses the API's `POST /v3/spaces/:guid/manifest_diff` endpoint,
which returns the diff in a JSON Patch format.  We use a recursive
algorithm to traverse the given manifest and print any differences
between it and the current state of the apps known to the API.

We changed the YAML struct tag for the disk quota field in the manifest
so that it defaults to `disk_quota` (with an underscore) rather than
`disk-quota` (with a hyphen). This means that the marshalled manifest
that we send to the API will have the underscore. We still accept
(unmarshal) this field with and without the underscore for backwards
compatibility.

[Finishes #173478933, Finishes #174425692]

Co-authored-by: Chris Selzo <cselzo@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>
Co-authored-by: Alexander Berezovsky <aberezovsky@pivotal.io>
Co-authored-by: Teal Stannard <tstannard@pivotal.io>