services: Move integration tests into v7 directory
Currently a number of services tests are in the "shared" directory which
dates from when the main branch was used for v6 and v7. But as this repo
now uses version branches, we should consoladate services tests into the
v7 directory.

Also removed the performance test. This was added some time ago as an
experiment, and the experiment has not been continued.

[#173836058](https://www.pivotaltracker.com/story/show/173836058)