CLI user can still apply a manifest if generating the diff 500s
When the API returns an unknown error (500) when we try to generate the
manifest diff, we just print a warning and move on. We still try to
apply the manifest. The goal is to be defensive against potential bugs
in this experimental API endpoint (we don't want diffing bugs to block
the intended workflow).

[finishes #174653498]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>