V3: default to large page size for list requests
When getting a large number of resources for a list endpoint, we
currently get 50 resources per page. The V3 maximum is 5000, so we could
make 100 times fewer HTTP requests which significantly improves
performance. This change means that we will request 5000 resources per
page except for situations where the code specifically requests a
different number.

[#175715881](https://www.pivotaltracker.com/story/show/175715881)