Revert "V3: default to large page size for list requests (#2109)"
This reverts commit fd7bde4d18cea16f8bb4b0b5ef71a1b92f2a1658.

The effects of this change were too widespread to be comfortable with.
It makes sense to only request the larger page size where needed, as
doing it by default seems to require a large number of test changes, and
makes those tests more brittle - for instannce if we wanted to change
the default in the future.