Explicitly set output filename for rsrc tool
We use this tool as part of building binaries for each platform:
https://github.com/akavel/rsrc

We install it with a `go get` in our Makefile. In CI, building these
binaries started failing with errors like `File 'rsrc.syso' not found`.

It turns out that a recent change to rsrc changed the default name of
the file that gets generated. It used to always default to `rsrc.syso`,
but now it defaults to something like `rsrc_{OS}_{arch}.syso`.

Fixing this by just explicitly passing `-o rsrc.syso` to rsrc, which
will preserve the behavior we were already relying on: creating a file
with the name `rsrc.syso`.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>